Refactored the code that uses flags to display/clear modals to use a struct so that I can pass that as a parameter instead of all of the individual fields, started work on matching orders

+++ src/Utils.cpp
@@ -17,7 +17,7 @@ bool isEmptyOrWhitespace(const std::string& str)
   return trim(str).empty();
 }
 
-u32 getCurrentTime()
+u64 getCurrentTime()
 {
   auto time = std::chrono::system_clock::now();
   auto dur = time.time_since_epoch();
